requirements.txt
# Core libraries 
numpy=1.24.4
pandas>=2.0.3
torch>=2.0.0
scikit-learn>=1.2.0

# Additional libraries
tqdm>=4.66.4
PyYAML>=6.0
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
albumentations>=1.4.8
glob2>=0.7
matplotlib>=3.7.5
matplotlib-inline>=0.1.7
argparse>=1.4.0

# Jupyter notebook libraries
ipykernel>=6.29.4
ipython>=8.12.3